# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "CV Processor API",
    "description": "API for processing CVs and evaluating them for O1A visa categories",
    "version": "1.0.0"
  },
  "paths": {
    "/process_cv/": {
      "post": {
        "summary": "Process CV",
        "description": "Upload a CV in PDF format for processing and O1A visa category evaluation",
        "operationId": "process_cv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CV file in PDF format"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "education": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "school": {
                            "type": "string"
                          },
                          "year": {
                            "type": "integer"
                          },
                          "degree": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "category_ratings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "category": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "string",
                            "enum": ["low", "medium", "high"]
                          },
                          "justification": {
                            "type": "string"
                          },
                          "information_used": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "information_unused": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}